windowFeature_api

+++ src/main/java/com/koala/learn/service/FileService.java
@@ -107,7 +107,9 @@ public class FileService {
         }else{
             File[] files = file.listFiles();
             for(int i=0;i<files.length;i++){
-                files[i].delete();
+                if(files[i].getName().startsWith("out")){
+                    files[i].delete();
+                }
             }
         }
     }
